Move filer-input from utils to ViewTrips.
